:bug: fix(modals): Fixing some bugs while modal was showing info
Now claim rewards dissapears if the rewards was already claimed
