+++ src/components/ModalDialogs/ModalGym.tsx
@@ -6,37 +6,67 @@ import { CustomButton } from '../CustomButton';
 import { ModalRewards } from './ModalRewards';
 import { GymsModalContext } from '@src/context/GymsModalContext';
 import { requestGymInfoById } from '@src/services/map.services';
-import { TGymInfo, TGymLoomieProtector } from '../../types/types';
+import { TGymInfo, TGymLoomieProtector, TReward } from '../../types/types';
+import { getPosition } from '@src/services/geolocation.services';
+import { requestRewards } from '@src/services/map.services';
 
-// todo cambiar nombres
 export const ModalGym = () => {
   const { isGymModalOpen, currentModalGymId, toggleGymModalVisibility } =
     useContext(GymsModalContext);
 
   const [gymInfo, setGymInfo] = useState<TGymInfo | null>();
+  const [reward, setReward] = useState(Array<TReward>);
+
+  // Request the gym infromation to show the first modal
   const fetchGymInfo = async () => {
     const [response, err] = await requestGymInfoById(currentModalGymId);
     if (!err) {
       const gymInfo = response;
       setGymInfo(gymInfo || null);
+      if (!gymInfo) return;
+    }
+  };
+
+  // Try to claim the gym rewards to show the second modal
+  const fetchClaimRewards = async () => {
+    const position = await getPosition();
+
+    if (position) {
+      const [response, err] = await requestRewards(
+        currentModalGymId,
+        position.lat,
+        position.lon
+      );
+
+      if (!err && response != null) {
+        const { reward } = response;
+        setReward(reward || null);
+      }
     }
   };
 
   useEffect(() => {
+    handleOpenRewardsModal();
+  }, [reward]);
+
+  useEffect(() => {
     if (currentModalGymId) {
-      //console.log({ currentModalGymId });
       fetchGymInfo();
     }
   }, [currentModalGymId]);
 
-  const [secondModalVisible, setSecondModalVisible] = useState(false);
+  const [rewardsModalVisible, setRewardsModalVisible] = useState(false);
 
-  const handleOpenSecondModal = () => {
-    setSecondModalVisible(true);
+  const handleOpenRewardsModal = () => {
+    setRewardsModalVisible(true);
   };
 
+  // This function is only called from the second modal to avoid changing
+  // the value of the was_reward_claimed variable.
   const handleCloseSecondModal = () => {
-    setSecondModalVisible(false);
+    setRewardsModalVisible(false);
+    if (!gymInfo) return;
+    gymInfo.was_reward_claimed = true;
   };
 
   const renderItem = ({ item }: { item: TGymLoomieProtector }) => (
@@ -71,11 +101,13 @@ export const ModalGym = () => {
             keyExtractor={(loomie) => loomie._id.toString()}
           />
           <View style={Styles.containerButton}>
-            <CustomButton
-              title='Claim Rewards'
-              type='primary'
-              callback={handleOpenSecondModal}
-            />
+            {!gymInfo?.was_reward_claimed && (
+              <CustomButton
+                title='Claim Rewards'
+                type='primary'
+                callback={fetchClaimRewards}
+              />
+            )}
             <CustomButton
               title='Challenge'
               type='primary'
@@ -87,8 +119,8 @@ export const ModalGym = () => {
         </View>
       </View>
       <ModalRewards
-        isVisible={secondModalVisible}
-        isClaimed={gymInfo?.was_reward_claimed || false}
+        isVisible={rewardsModalVisible}
+        reward={reward}
         callBack={handleCloseSecondModal}
       />
     </Modal>
@@ -103,7 +135,7 @@ const Styles = StyleSheet.create({
   },
   modal: {
     backgroundColor: '#fff',
-    maxHeight: '52%',
+    maxHeight: 375,
     width: '96%',
     padding: 12
   },
