+++ src/services/map.services.ts
@@ -138,7 +138,6 @@ export const requestRewards = async (
         }
       }
     );
-    console.log(response.data);
 
     return [response.data, false];
   } catch (error) {
